+++ data-raw/ex_fit.R
@@ -6,7 +6,6 @@ segments = list(
 )
 
 # Fit it. The `ex_demo` dataset is included in mcp
-options(mc.cores = 3)
 ex_fit = mcp(segments, data = ex_demo, adapt = 3000, iter = 1000, sample = "both")
 ex_fit$mcmc_loglik = NULL  # Make the object small